+++ vast_pipeline/pipeline/new_sources.py
@@ -12,9 +12,8 @@ from astropy.wcs.utils import (
 )
 
 from vast_pipeline.models import Image, Run
-from vast_pipeline.utils.utils import StopWatch
+from vast_pipeline.utils.utils import StopWatch, calculate_n_partitions
 from vast_pipeline.image.utils import open_fits
-from vast_pipeline.pipeline.utils import calculate_n_partitions
 
 
 logger = logging.getLogger(__name__)
