+++ vast_pipeline/pipeline/pairs.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 from psutil import cpu_count
 
-from vast_pipeline.pipeline.utils import calculate_n_partitions
+from vast_pipeline.utils.utils import calculate_n_partitions
 
 
 logger = logging.getLogger(__name__)
