+++ vast_pipeline/pipeline/utils.py
@@ -24,7 +24,8 @@ from itertools import chain
 from vast_pipeline.image.main import FitsImage, SelavyImage
 from vast_pipeline.image.utils import open_fits
 from vast_pipeline.utils.utils import (
-    eq_to_cart, StopWatch, optimize_ints, optimize_floats
+    eq_to_cart, StopWatch, optimize_ints, optimize_floats,
+    calculate_n_partitions
 )
 from vast_pipeline.models import (
     Band, Image, Run, SkyRegion
@@ -1713,26 +1714,3 @@ def write_parquets(
     )
 
     return skyregs_df
-    
-def calculate_n_partitions(df, n_cpu, partition_size_mb=100):
-    """
-    This function will calculate how many partitions a dataframe should be
-    split into.
-    
-    Args:
-        df: The pandas dataframe to be partitionined.
-        n_cpu: The number of available CPUs.
-        partition_size: The optimal partition size in MB.
-    Returns:
-        The optimal number of partitions.
-    """
-    mem_usage_mb = df.memory_usage(deep=True).sum() / 1e6
-    n_partitions = int(np.ceil(mem_usage_mb/partition_size_mb))
-    
-    # n_partitions should be >= n_cpu for optimal parallel processing
-    if n_partitions < n_cpu:
-        n_partitions=n_cpu
-    
-    logger.debug("Using {n_partitions} partions of {partition_size}MB")
-
-    return n_partitions
